[move/#8] Genre파일 위치 변경
