+++ core/designsystem/src/main/java/com/napzak/market/designsystem/component/bottomsheet/GenreSearchBottomSheet.kt
@@ -33,7 +33,6 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.napzak.market.common.type.Genre
 import com.napzak.market.designsystem.R.drawable.ic_x_button_24
 import com.napzak.market.designsystem.R.string.genre_search_hint
 import com.napzak.market.designsystem.R.string.genre_search_select_genre
